Remove build_viz workflow from FactIQ skill
